requirements.txt
Flask==3.0.2
gunicorn==21.2.0
gevent==23.9.1

redis==5.0.1
psycopg2-binary==2.9.9

python-dotenv==1.0.1
Werkzeug==3.0.1
